Make get_wstp_link() private
If users want to call getWSLINK(), the method is still fully exposed on
WolframLibraryData, with a bit of `unsafe` code.

This method may be made public again in the future, it's just that exposing use
of WSTP under the hood is an implementation detail which we may want to change
in the future.