bugfix: disable generation of bindings using WolframCompileLibrary.h
WolframCompileLibrary.h provides a concrete definition of `struct st_MNumericArray`,
but includes a big warning at the top of the file that the definitions from that
header are not guaranteed to work from one version of the system to another. st_MNumericArray
should be an opaque type which is only handled via a pointer and callback functions.