bugfix: Replace `WolframEngine::get_wstp_link() -> Link` with `with_link(FnMut(&mut Link) -> R)` free function
get_wstp_link() was conceptually wrong because it would take ownership of the value
returned from `getWSLINK()` (and the link would be closed when the Link was dropped),
when really getWSLINK() returns a borrowed value.

Use a Mutex to try to enforce exclusive access to the getWSLINK() value from safe
code.