bugfix: Add assert_main_thread() helper to ensure thread safety in wolfram-library-link's safe API
Not all Kernel API's can safely be called from any thread other than the main Kernel thread.
To attempt to minimize the possibility of that happening, this commit adds a new
assert_main_thread() function that will panic if the current thread is not the main Kernel
thread.

Over time, this function should be used in more places inside of wolfram-library-link, or
intentionally not used for API's that only call RTL functions that are safe to use from
non-main threads.

* Make `initialize()` unsafe. See Safety comment for details.
* Add examples/tests/test_threading.rs and Tests/Threading.wlt