bugfix: correctly capture messages from PanicInfo payloads from panic!'s that use format arguments + add test
For panics that don't use any format arguments, the PanicInfo::payload is always
a &str ref, a reference to the static str that was the first argument of the panic!
invocation.

However, for panic! invocations that used format arguments, a String has to be
allocated to hold the final formatted message.