feature: Support `fn foo(args: Vec<Expr>) -> Expr` functions exported via export_wstp![]
This tool longer than you might expect, and required changing the syntax of the
export_wstp![] macro, due to some slightly odd behavior when using type
inference and casting a `fn(...) {some_name}` type to `fn(...)`. This took quiet
a while to sort out exactly what was going wrong, and leads to the slight
inconvenience that when using the export_wstp! macro with a `fn foo(&mut Link)`
function, you must specify the `&mut _` portion of the type explicitly in the
export_wstp! invocation, like so:

	export_wstp![foo(&mut _)]

in the examples I've just written the full type out:

	export_wstp![foo(&mut Link)]

which I think reads more clearly than the half-inferred half-concrete
oddity that `&mut _` is.

Change list:

* Add WstpFunction trait
  - implemented for:
    * fn(&mut Link)
    * fn(Vec<Expr>) -> Expr
    * fn(Vec<Expr>)
* Add total(args: Vec<Expr>) example