Add FIXME for unsoundness issue with export_wstp! when using with a `fn(&'static mut _)` function
This should be a type error. Fixing this will require changing the declaration
of `F` in call_native_wolfram_library_function() to `for<'a> NativeFunction<'a>`,
so that the 'a lifetime is chosen by the implementation of
call_native_wolfram_library_function(), and NOT by the caller.

Doing so is not completely trivial, and will require propagation of that
change, like what happened in an earlier commit with export_wstp!