docs: Rename 'Call back into Wolfram from Rust' to 'Evaluate Wolfram code from Rust'
I changed my mind. The phrase 'call back' is a bit more technical and less
likely to be familiar than simply 'evaluate', which is a clearer statement of
what's happening.

Additionally, a 'callback function' (i.e. a function pointer / closure)
is one mechanism for executing code across a library / language boundary, but that
is not at all the mechanism used for Rust -> Wolfram evaluations. The new name
avoids this potentially misleading connotation.