Adds u8, u16, and u32 support to DataStore
Extends the `DataStore` functionality to support `u8`, `u16`, and `u32` types.

This change introduces implementations of the `DataStoreAdd` trait for these unsigned integer types, allowing them to be seamlessly added to a `DataStore`. Corresponding test cases and Wolfram Language bindings are included to verify the correctness of these new features.